% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Earthquake data cleaning}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{A dataframe of the NOAA earthquake dataset}
}
\value{
A dataframe.
}
\description{
This function can be used to clean the NOAA earthquake dataset. E. g. it
formats the data correctly.
}
