% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Earthquake label creation}
\usage{
eq_create_label(x)
}
\arguments{
\item{x}{A dataframe containing the cleaned NOAA earthquake dataset.}
}
\value{
A character vector
}
\description{
This function creates a HTML-formatted earthquake description containing the
location, magnitude and death count. It was built to serve as the label text
for the function \code{\link[=eq_map]{eq_map()}}.
}
\examples{
eq_create_label(noaa_earthquakes)
 
}
