% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Earthquake location name cleaning}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{A dataframe containing the cleaned NOAA earthquake dataset.}
}
\value{
A dataframe.
}
\description{
This function extracts the country name from the location name column and
further cleans the location name column.
}
\examples{
eq_location_clean(noaa_earthquakes)

}
