% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake map}
\usage{
eq_map(x, annot_col)
}
\arguments{
\item{x}{A dataframe containing the cleaned NOAA earthquake dataset.}

\item{annot_col}{A column name of the given dataframe to be used as label text.}
}
\value{
A leaflet map.
}
\description{
This function creates a leaflet map visualizing the earthquake locations.
}
\examples{
eq_map(noaa_earthquakes, "location_name")

}
