% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  xmin = as.Date("0001-01-01"),
  xmax = as.Date("9999-01-01"),
  n_max = 5,
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{xmin}{A date value to filter the data from the given date.}

\item{xmax}{A date value to filter the data up to the given date.}

\item{n_max}{Maximum amount of earthquakes that will be labelled. The
earthquakes with the greatest size value are labelled first.}
}
\value{
A ggplot graph.
}
\description{
This geom labels the earthquakes that have been visualized by the
\code{\link[=geom_timeline]{geom_timeline()}} function with the location name.
}
