% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa_earthquakes}
\alias{noaa_earthquakes}
\title{National Oceanographic and Atmospheric Administration (NOAA) earthquake data}
\format{
\subsection{\code{noaa_earthquakes}}{

A data frame with 6478 rows and 38 columns:
\describe{
\item{year}{Year}
\item{mo}{Month}
\item{dy}{Day}
\item{location_name}{Country and location name}
\item{latitude}{Latitude}
\item{longitude}{Longitude}
\item{mag}{Magnitude}
\item{total_deaths}{Death count}
...
}
}
}
\source{
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}
}
\usage{
noaa_earthquakes
}
\description{
The Significant Earthquake Database contains information on destructive
earthquakes from 2150 B.C. to 2025 that meet at least one of the
following criteria: Moderate damage (approximately $1 million or more),
10 or more deaths, Magnitude 7.5 or greater, Modified Mercalli Intensity
X or greater, or the earthquake generated a tsunami.
}
\keyword{datasets}
